removing fee analytics as a service, making it a helper
